LPS-195481 Inline CertificateUploadFileEntryHandler in UpdateCertificateMVCResourceCommand
